+++ src/main/java/luminous/safechat/util/Fmt.java
@@ -1,0 +1,17 @@
+package luminous.safechat.util;
+
+import org.jetbrains.annotations.NotNull;
+
+import java.util.Map;
+
+public class Fmt {
+    public static String format(@NotNull String fmt, @NotNull Map<@NotNull String, Object> args) {
+        String result = fmt;
+        for (var arg : args.entrySet()) {
+            if (arg.getValue() == null) continue;
+
+            result = result.replaceAll("\\{" + arg.getKey() + "}", arg.getValue().toString());
+        }
+        return result;
+    }
+}